WIP: Use Abort marker trait to audit OOM panics
This avoids the non-untuitive behavior of the `try_*` methods from
`Error = !`, while still giving us the ability to audit panicks.
Perfect!

Credit to @TimDiekmann for the idea.

`AbortAlloc` is put back as an private implementation detail to make the
unchecked unwrapps in `boxed` safe. A simpler solution would be just get
in the habit of returning the layout and allocation error, as in the
general case the caller has now idea which allocation deep in the callee
failed. Not returning the layout only makes sense for the allocator
trait itself, when the caller must know the layout.

WIP because there are some doc test compile failures I don't have time
to sort out at this moment.